Simplify implementaton
